included bootstrap in bower added prompt for deletion.

+++ app/controllers/users.js
@@ -3,8 +3,12 @@ import Ember from 'ember';
 export default Ember.Controller.extend({
 	 actions : {
         remove : function(user){
-            user.deleteRecord();
-            user.save();
+        	 var x = confirm("Are you sure you want to delete?");
+             if(x)
+             {
+          	  user.deleteRecord();
+              user.save();
+             }    
+           }         
         }
-    }
-});
+    });
